requirements.txt
flask==3.0.3
gunicorn==22.0.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
deepface==0.0.93
tf-keras==2.16.0
mediapipe==0.10.14
scikit-image==0.24.0
pillow==10.4.0
scipy==1.13.1